Unify schema and service into a Interface I guess
