+++ src/db/schemas/index.js
@@ -1,6 +1,5 @@
-import Task from './Task.js'
 import Template from './Template.js'
 import User from './User.js'
 import Icon from './Icon.js'
 
-export { Task, Template, User, Icon }
+export { Template, User, Icon }
