+++ src/db/task-service.js
@@ -1,82 +1,11 @@
-import { deleteImage } from './helpers.js'
-import { get } from 'svelte/store'
-import { user, tasksCache } from '/src/store/index.js'
-import { Task } from '../db/schemas'
-import { writeBatch } from 'firebase/firestore'
-import { db } from './init'
-import { 
-  maintainTreeISOs, 
-  maintainTreeISOsForCreate,
-  handleTreeISOsForDeletion
-} from './treeISOs.js'
-import { doc } from 'firebase/firestore'
+import Task from './Task.js'
 
-export async function createTaskNode ({ id, newTaskObj }) {
-  try {
-    const batch = writeBatch(db)
-    const validatedTask = Task.parse({ ...newTaskObj })
-    const treeISOs = maintainTreeISOsForCreate({ task: validatedTask, batch })
-    
-    batch.set(doc(db, `users/${get(user).uid}/tasks/${id}`), { 
-      treeISOs,
-      ...validatedTask
-    })
+// Re-export with original names for backwards compatibility
+export const createTaskNode = (params) => Task.create(params)
+export const updateTaskNode = (params) => Task.update(params)
+export const deleteTaskNode = (params) => Task.delete(params)
 
-    batch.commit()
-  } 
-  catch (error) {
-    console.error('Error creating task:', error)
-    alert('Error creating task: ' + error.message)
-    return error
-  }
-}
-
-export async function updateTaskNode ({ id, keyValueChanges }) {
-  try {
-    const validatedChanges = Task.partial().parse(keyValueChanges)
-    const batch = writeBatch(db)
-    maintainTreeISOs({ id, keyValueChanges: validatedChanges, batch })
-    batch.update(doc(db, `users/${get(user).uid}/tasks/${id}`), validatedChanges)
-
-    batch.commit()
-  }
-  catch (error) {
-    alert("Error saving changes to db, please reload");
-    console.error("Error in updateTaskNode: ", error);
-  }
-}
-
-export async function deleteTaskNode ({ id }) {
-  const task = get(tasksCache)[id]
-
-  const tasksToDelete = [task]
-  pushDescendants(task.id, tasksToDelete)
-
-  if (tasksToDelete.length >= 2) {
-    if (!confirm(`${tasksToDelete.length} tasks will be deleted in this tree. Are you sure?`)) {
-      return
-    }
-  }
-
-  const { uid } = get(user)
-  const batch = writeBatch(db)
-
-  for (const task of tasksToDelete) {
-    const { imageFullPath, id } = task
-    if (imageFullPath) deleteImage({ imageFullPath })
-    batch.delete(doc(db, `/users/${uid}/tasks/${id}`))
-  }
-  handleTreeISOsForDeletion({ tasksToDelete, batch })
-
-  await batch.commit()
-  return tasksToDelete
-}
-
-function pushDescendants (parentID, tasksToDelete) {
-  Object.values(get(tasksCache)).forEach(t => {
-    if (t.parentID === parentID) {
-      tasksToDelete.push(t)
-      pushDescendants(t.id, tasksToDelete)
-    }
-  })
+// Add deprecation warnings in development
+if (process.env.NODE_ENV !== 'production') {
+  console.warn('task-service.js is deprecated. Import from /src/db/Task.js instead.')
 }
