+++ src/lib/Reusable/PhotoTaskElement.svelte
@@ -99,7 +99,7 @@
  import { grabOffset, activeDragItem, openTaskPopup } from '/src/store'
  import { lazyCallable } from '/src/utils/svelteActions.js'
  import { pixelsPerHour } from '/src/lib/Calendar/store.js'
- import { updateTaskNode } from '/src/db/task-service.js'
+ import Task from '/src/db/Task.js'
 
  export let task = null
 
@@ -142,7 +142,7 @@
    const newY = getTrueY(e)
    const durationChange = minutesPerPixel * (newY - startY)
 
-   updateTaskNode({
+   Task.update({
      id: task.id,
      keyValueChanges: {
        duration: Math.max(1, task.duration + durationChange) // can't have a 0 duration event
